+++ json_serializable/test/default_value/default_value.checked.g.dart
@@ -47,8 +47,7 @@ DefaultValue _$DefaultValueFromJson(Map json) {
         json,
         'fieldEnum',
         (v) => val.fieldEnum =
-            $enumDecodeNullable('Greek', Greek.values, v as String) ??
-                Greek.beta);
+            _$enumDecodeNullable(_$GreekEnumMap, v) ?? Greek.beta);
     return val;
   });
 }
@@ -72,6 +71,33 @@ Map<String, dynamic> _$DefaultValueToJson(DefaultValue instance) {
   val['fieldListSimple'] = instance.fieldListSimple;
   val['fieldMapSimple'] = instance.fieldMapSimple;
   val['fieldMapListString'] = instance.fieldMapListString;
-  val['fieldEnum'] = instance.fieldEnum?.toString()?.split('.')?.last;
+  val['fieldEnum'] = _$GreekEnumMap[instance.fieldEnum];
   return val;
 }
+
+T _$enumDecode<T>(Map<T, dynamic> enumValues, dynamic source) {
+  if (source == null) {
+    throw new ArgumentError('A value must be provided. Supported values: '
+        '${enumValues.values.join(', ')}');
+  }
+  return enumValues.entries
+      .singleWhere((e) => e.value == source,
+          orElse: () => throw new ArgumentError(
+              '`$source` is not one of the supported values: '
+              '${enumValues.values.join(', ')}'))
+      .key;
+}
+
+T _$enumDecodeNullable<T>(Map<T, dynamic> enumValues, dynamic source) {
+  if (source == null) {
+    return null;
+  }
+  return _$enumDecode<T>(enumValues, source);
+}
+
+const _$GreekEnumMap = const <Greek, dynamic>{
+  Greek.alpha: 'alpha',
+  Greek.beta: 'beta',
+  Greek.gamma: 'gamma',
+  Greek.delta: 'delta'
+};